🎉 clean definition of interfaces
